ref: separate growth models in their own files
