+++ forestry/grow_motti.py
@@ -1,0 +1,5 @@
+from forestry.ForestDataModels import ForestStand
+
+
+def grow_motti(stand: ForestStand, **operation_params) -> ForestStand:
+    raise NotImplementedError("TODO")
